requirements.txt
django==5.2.2
djangorestframework
django-rest-knox
markdown
django-filter
gunicorn==23.0.0
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
requests==2.32.3
freezegun==1.5.2